Rename bayesian_jobs/ to f8a_jobs/
