+++ Dockerfile
@@ -27,7 +27,7 @@ COPY hack/update_selinon.sh /tmp/
 RUN sh /tmp/update_selinon.sh
 
 COPY hack/run_jobs.sh /usr/bin/
-COPY bayesian-jobs.py /usr/bin/
+COPY f8a-jobs.py /usr/bin/
 
 USER coreapi
 CMD ["/usr/bin/run_jobs.sh"]
