+++ hack/run_jobs.sh
@@ -5,6 +5,6 @@ set -ex
 DIR=$(dirname "${BASH_SOURCE[0]}")
 source $DIR/env.sh
 
-bayesian-jobs.py initjobs
+f8a-jobs.py initjobs
 cd /usr/bin/
-exec uwsgi --http 0.0.0.0:34000 -p 1 -w bayesian-jobs --enable-threads
+exec uwsgi --http 0.0.0.0:34000 -p 1 -w f8a-jobs --enable-threads
